# docs/swagger.yaml
basePath: /api/v1/
definitions:
  domain.Ad:
    properties:
      category:
        type: string
      contacts:
        type: string
      description:
        type: string
      id:
        type: integer
      imagesURL:
        items:
          type: string
        type: array
      price:
        type: integer
      published:
        type: boolean
      title:
        type: string
      userId:
        type: string
    type: object
  repository.FtsResponse:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
  v1.adminSignInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  v1.inputAd:
    properties:
      category:
        type: string
      contacts:
        $ref: '#/definitions/v1.inputContacts'
      description:
        type: string
      images_url:
        items:
          type: string
        type: array
      price:
        type: integer
      published:
        type: boolean
      title:
        type: string
    required:
    - category
    - contacts
    - description
    - images_url
    - price
    - title
    type: object
  v1.inputContacts:
    properties:
      email:
        type: string
      location:
        type: string
      name:
        type: string
      phone_number:
        type: string
    required:
    - email
    - location
    - name
    - phone_number
    type: object
  v1.inputFTSRequest:
    properties:
      request:
        type: string
    required:
    - request
    type: object
  v1.refreshTokensInput:
    properties:
      RefreshToken:
        type: string
    required:
    - RefreshToken
    type: object
  v1.response:
    properties:
      message:
        type: string
    type: object
  v1.signInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  v1.signUpInput:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
  v1.tokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for Posting Ads
  title: Application for posting ads
  version: "1.0"
paths:
  /admins/Sign-In:
    post:
      consumes:
      - application/json
      description: admin sign in
      parameters:
      - description: sign in info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.adminSignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Admin SignIn
      tags:
      - admin-auth
  /admins/api/ads/:
    get:
      consumes:
      - application/json
      description: admin get all users ads
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Ad'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get All His Ads
      tags:
      - admin-ads
  /admins/api/ads/{id}:
    delete:
      consumes:
      - application/json
      description: admin delete user ad by adId
      parameters:
      - description: adId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Delete Ad
      tags:
      - admin-ads
    get:
      consumes:
      - application/json
      description: admin get ad by adId
      parameters:
      - description: adId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get Ad By AdId
      tags:
      - admin-ads
    put:
      consumes:
      - application/json
      description: admin create ad
      parameters:
      - description: adId
        in: path
        name: id
        required: true
        type: string
      - description: ad info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.inputAd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Update User Ad
      tags:
      - admin-ads
  /admins/refreshTokens:
    post:
      consumes:
      - application/json
      description: admin refresh tokens
      parameters:
      - description: refresh token info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.refreshTokensInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Admin Refresh Tokens
      tags:
      - admin-auth
  /auth/Sign-In:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: sign in info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User SignIn
      tags:
      - users-auth
  /auth/Sign-Up:
    post:
      consumes:
      - application/json
      description: create user account
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User SignUp
      tags:
      - users-auth
  /auth/api/ads/:
    get:
      consumes:
      - application/json
      description: user get all his ads by userId
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Ad'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Get All His Ads
      tags:
      - users-ads
    post:
      consumes:
      - application/json
      description: user create his own ad
      parameters:
      - description: create ad info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.inputAd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Create Own Ad
      tags:
      - users-ads
  /auth/api/ads/{id}:
    delete:
      consumes:
      - application/json
      description: user delete his ad
      parameters:
      - description: adId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Delete Ad
      tags:
      - users-ads
    get:
      consumes:
      - application/json
      description: user get ad by adId
      parameters:
      - description: adId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Get Ad By AdId
      tags:
      - users-ads
    put:
      consumes:
      - application/json
      description: user create ad
      parameters:
      - description: adId
        in: path
        name: id
        required: true
        type: string
      - description: ad info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.inputAd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Update His Ad
      tags:
      - users-ads
  /auth/api/fts/:
    get:
      consumes:
      - application/json
      description: user search ads by his request string
      parameters:
      - description: search request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.inputFTSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.FtsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - UsersAuth: []
      summary: User Search Ads
      tags:
      - users-ads
  /auth/refreshTokens:
    post:
      consumes:
      - application/json
      description: user refresh tokens
      parameters:
      - description: refresh token info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.refreshTokensInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User Refresh Tokens
      tags:
      - users-auth
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  UsersAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"